Conclusão da primeira fase do lançamento de notas

+++ src/main/java/br/com/cng12/clubdance/entity/ProdutoEntity.java
@@ -32,7 +32,7 @@ public class ProdutoEntity {
 	private Double preco;
 	
 	@Column(nullable = false)
-	private Double qtdeEstoque;
+	private int qtdeEstoque;
 	
 	@Column(nullable = true)
 	private String unidadeMedida;
